@EventListener Logic moved to new file - OnStartup
