+++ src/main/java/it/eng/idsa/businesslogic/util/AutoSelfRegistration.java
@@ -9,8 +9,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.context.event.ApplicationReadyEvent;
-import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 
 import de.fraunhofer.iais.eis.util.ConstraintViolationException;
@@ -20,7 +18,7 @@ import it.eng.idsa.businesslogic.service.SelfDescriptionService;
 
 @ConditionalOnProperty(name="application.selfdescription.registrateOnStartup", havingValue="true")
 @Component
-public class AutoSelfRegistration{
+public class AutoSelfRegistration {
 	
 	private static final Logger logger = LoggerFactory.getLogger(AutoSelfRegistration.class);
 	
@@ -30,16 +28,6 @@ public class AutoSelfRegistration{
 	@Autowired
 	private BrokerService brokerService;
 	
-	@EventListener(ApplicationReadyEvent.class)
-	public void selfRegistrate() throws ConstraintViolationException, URISyntaxException, DatatypeConfigurationException {
-		logger.info("Starting AutoSelfRegistration");
-
-		brokerService.sendBrokerRequest(selfDescriptionService.getConnectorAvailbilityMessage(),
-				selfDescriptionService.getConnectorSelfDescription());
-
-		logger.info("AutoSelfRegistration finished");
-	}
-	
 	@PreDestroy
 	public void selfPassivate() throws ConstraintViolationException, URISyntaxException, DatatypeConfigurationException {
 		logger.info("Starting sign out process from broker");
@@ -48,9 +36,5 @@ public class AutoSelfRegistration{
 				selfDescriptionService.getConnectorSelfDescription());
 
 		logger.info("Sign out process finished");
-
 	}
-	
-
-
 }
